HOLY COW I FOUND THE BUG !!
Explanation:
When I was passing the AU radius props for both target and observer,
I needed to convert them into pixel coordinates. Before I started using AU,
I used getPlanetPos(radius, phase) with pixels, and did it that way, but after
making the switch to AU, it resulted in round off errors acccumlating.
THIS IS SIMPLY A HYPOTHESIS (for now), but Im pretty sure the cause of the
bug was caused mathematically by the separation of logic with AU and pixel.
I need to investigate the code for more possible lurking bugs.