Put correct zodiac img, added styling, format text
The bug written about in my previous report was actually
coming from a separate location (not math round off errors).
It's happening because pixel radius will always be 400 for
the outer planet, therefore any change to AU will not be
reflected in the planet's pixel radius. I also realized
that the sprite object must be stored in order to change
a planet image.